tests: Setup Python path on runner module import
Previously the Python path was being set up as part ofrunning the
runner.standalone() helper on a module. That's unintuitive because it
means that local modules can only be imported in the test classes
themselves rather than at the top of the test module.

Making the Python path setup part of the runner module import, local
modules can now be imported in a more natural way. The only restriction
is that they be imported after the runner module.

Signed-off-by: Thierry Reding <treding@nvidia.com>