boards: Add 'urandom_read' warning to whitelist
During boot messages such as the following maybe seen ...

     WARNING KERN urandom_read: 6 callbacks suppressed

These may occur if there is not enough entropy available when urandom
is read. Although this should be fixed by adding more entropy sources,
given that this has always occurred when booting upstream Linux kernels
on Tegra, whitelist this for now.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>