tests: cpu: cpufreq requires userspace governor
The cpufreq test needs the userspace governor to be available in order
to step through the various CPU frequencies. Make sure to skip the test
if that userspace governor is not available.

Note that ideally we'd just skip this test entirely using one of the
skip decorators. However there's some use in running at least part of
the test, which will list a bunch of stuff from sysfs that might be
interesting to keep track of. Also, technically each CPU can support a
different set of scaling governors, so we can't skip the test wholesale
if one of the CPUs doesn't support the userspace governor.

Signed-off-by: Thierry Reding <treding@nvidia.com>