tests: memory: EMC test requires frequency scaling support
The EMC test requires the kernel to have enabled EMC frequency support.
This is currently determined by the presence of the supported_rates file
in debugfs, which is how this support is exposed on Tegra K1. For newer
chips this may have to be adjusted based on what the implementation
provides.

Signed-off-by: Thierry Reding <treding@nvidia.com>