Add debugging terminal tool
This terminal tool can be used to automate various tasks on a running
system. By default it will try to find a PM342 board and connect to the
ttyUSB that's normally connected to the device's debug UART. It will
then start to monitor the data stream from the debug UART and run some
commands.

Note that this is not polished yet. The commands that are run are mostly
arbitrary. Eventually this is meant to allow separate test runs to be
easily added. One potentially interesting use-case for this is to allow
running it as part of an automated regression test using git bisect.
Typically the bisection will have to build various revisions and then
the user is expected to report test results. If properly written, these
tegra-term tests could serve as that step to fully automate bisection.

Signed-off-by: Thierry Reding <treding@nvidia.com>