linux: sysfs: Allow devices to be created from path or bus/name
In some use-cases it is practical to create a device from a full path
rather than from a bus/name pair. Allow callers to specify either the
full path or a bus/name pair. Also allow a path/name pair to create a
device given a partial path (to the device's parent) and the name of
the device.

Signed-off-by: Thierry Reding <treding@nvidia.com>