sysfs: Allow multiple drivers for devices
Depending on the kernel version or the firmware interface, a given
device may bind to a different driver. One such example is the UART
which will bind to the of_serial driver on older kernels, but with
v5.7 and later, a Tegra-specific driver was added so that it can
also support ACPI. The new driver is tegra-uart. List both of them
as a list of drivers and extend the test to iterate through that
list in order to match.

Signed-off-by: Thierry Reding <treding@nvidia.com>