linux: Support linux-next kernel versions
linux-next is an integration branch and its contents represent what will
usually go into the next release. In order to support versioned checks
on linux-next, add code to identify linux-next releases and special-case
the comparison against a literal kernel version.

This allows to versioned checks for the next release (say v5.9) to
succeed on a linux-next of the current release (based on v5.8-rcX).

Signed-off-by: Thierry Reding <treding@nvidia.com>