tegra-tests: Ignore blank lines in kernel log
Some kernel warnings reported contain only a newline character and so
are not really proper warnings themselves. These warnings can occur
when a function calls 'pr_cont("\n")' to print a newline but the print
occurs after an unrelated warning message causing this newline to be
reported also as a warning. Therefore, skip any lines in the kernel
log that only contains a newline character.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>