tegra-tests: Ignore backlight for Jetson TX1
By default the Jetson TX1 developer kit does not include an LCD panel,
and if it is not present then kernel warnings are issued because the LCD
backlight device is not found. Add these warning messages to the allowed
list for Jetson TX1 and remove the LCD backlight device from the default
list of device for Jetson TX1.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>