boards: Add ASoC warning to allowlist for Tegra
Add the ASoC warning 'ASoC: no DMI vendor name!' to the allowlist for
Tegra platforms. This warning is seen on Linux kernels when registering
an ASoC sound-card with the Linux kernel for platforms that are not
using ACPI and hence, do not provide a DMI table. This warning will be
fixed starting with Linux kernel v5.13 but for earlier kernels this can
be added to the allowlist.

Signed-off-by: Thierry Reding <treding@nvidia.com>