board: Add PMU error message for allowed list for Tegra194 platforms
Linux v5.13 added a new error message that would be displayed on boot
if the PMU is not found.

	ERR KERN kvm: pmu event creation failed -2

The PMU node was populated for Tegra194 platforms and so this error
message would be seen on boot. The PMU node has been added for Linux
v5.14 but for v5.13 add this error to the allowed list.

Signed-off-by: Thierry Reding <treding@nvidia.com>