board: Add IRQ affinity warning to allowlist for Jetson Xavier
When entering suspend on Jetson Xavier, the following warning is
observed ...

 WARNING KERN IRQ77: set affinity failed(-22).

In this case, IRQ77 on Jetson Xavier is the PCI-MSI interrupt for the
ACHI controller and the above warning is displayed because it is not
possible to set the IRQ affinity for the interrupt. Ideally, the kernel
would not warn in this case, but unfortunately there is no easy way to
tell the kernel that setting the affinity is not supported. Therefore,
add the 'set affinity failed' to the allowlist for only the AHCI
interrupt.

Signed-off-by: Thierry Reding <treding@nvidia.com>