board: Update PMU error handling for Tegra194
Commit ae8d425a801d ("board: Add PMU error message for allowed list for
Tegra194 platforms") added a PMU warning to the allow-list for Tegra194
and Linux v5.13 kernels. However, this warning is allowed for v5.13.0
kernels only. This warning is still seen on stable 5.13 kernels such
as 5.13.2 and so rather than only allow for 5.13.0, allow this warning
for kernels that are older than 5.14.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>