board: Fix kernel version for Jetson Xavier NX audio devices
Per the comment in the Jetson Xavier NX board file, the audio devices
for Jetson Xavier were added starting with Linux kernel version v5.14.
However, the board file incorrectly checks for this audio devices on
v5.12+ kernels and this causes the device test to fail on v5.12 and
v5.13 kernels. Fix this by updating the kernel version for these audio
devices.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>