boards: Add PCIe warning to the allowed list
Since Linux v6.0, the following error message is observed with Jetson
Xavier on boot ...

 tegra194-pcie 14140000.pcie: Phy link never came up

This error message was added by Linux kernel commit 14c4ad125cf9 ("PCI:
dwc: Log link speed and width if it comes up") and is observed on Jetson
Xavier when there is no PCI device connected. This error message will be
changed to an info level message for Linux v6.2, but for prior kernel
versions add this to the allowed list.

Signed-off-by: Thierry Reding <treding@nvidia.com>