boards: Update memory-controller warnings
With Linux v6.5, the following warnings are observed on Tegra194
devices ...

 tegra186-emc 2c60000.external-memory-controller: sync_state() pending due to 17000000.gpu
 tegra186-emc 2c60000.external-memory-controller: sync_state() pending due to 15380000.nvjpg
 tegra186-emc 2c60000.external-memory-controller: sync_state() pending due to 154c0000.nvenc
 tegra186-emc 2c60000.external-memory-controller: sync_state() pending due to 15a80000.nvenc

Similar warnings were already added to the allowed-list for the tegra-mc
device with Linux v6.4 because there is no upstream driver for these
devices. For Linux v6.5 the above warnings are seen for the tegra186-emc
device and so add these new variants as well.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>