linux: Fix version comparisons for patch level greater than 255
When the Linux patch level is greater than 255, it overflows and
increments the minor version incorrectly. For example, when comparing
v5.4.256 against v5.5.0, it appears that v5.4.256 is greater than v5.5.0
which is incorrect. Fix this by allocating 16 bits for each level of the
kernel version string.

Finally, add a new 'code' helper that returns a version code that is
compatible with the LINUX_VERSION_CODE that is generated by the Linux
kernel.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>