boards: Update Reserved memory warning
In Linux v6.9-rc1, commit 54c180e73ffa ("of: Move all FDT
reserved-memory handling into of_reserved_mem.c") moved the device-tree
reserved memory code to another location and this changed the prefix in
the following kernel warning from 'OF: fdt:' to 'OF: reserved mem'.

 OF: reserved mem: Reserved memory: unsupported node format, ignoring

After this commit the above warning is now observed again for Linux v6.9.
The reason this warning is observed is because the Tegra CBoot
bootloader is populating a 'grid-of-semaphores' node under
'reserved-memory' but it not populating the '#size-cells' and
'#address-cells'.

The CBoot bootloader needs to be fixed to correct this, but given that
this is a long-standing issue that has been added to the allowed kernel
warning list, update the regex for the kernel warning to include the
new prefix. Note that this observed with the CBoot bootloader that is
distributed with the NVIDIA L4T r32 BSP for Jetson.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>