kmsg: Handle printk record extensions
The ABI mentions that one or more extensions can follow the standard
fields in the printk records and consumers should ignore any unknown
fields for forward compatibility.

Read extensions into a separate field so that they can be ignored.
Remove the message that was flagging these additional fields because
it causes a lot of noise if extensions are added.

This fixes an issue observed on SLES systems where the PRINTK_CALLER
Kconfig option is enabled, which causes the extra "caller" extension
to be added to every printk record.

Reported-by: Paritosh Dixit <paritoshd@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>