tests: Add an iperf test
This uses iperf3 to run a benchmark on the default interface, which will
allow checking for performance regressions. The test requires the
interface to perform at 90% of the theoretical maximum (as specified by
the current link speed).

Included as part of this is a new network oriented module that relies on
the "ip" utility (from iproute2) to retrieve the necessary data.

Signed-off-by: Thierry Reding <treding@nvidia.com>