linux: sysfs: Add equality operator for sysfs objects
The equality operator is useful to check whether two paths point to the
same object. sysfs has a lot of different paths (via subsystems, busses)
to reach a specific device, so this operator is useful to check for
equality between them. Among other things this can be used to remove
objects from containers, etc.

Suggested-by: Paritosh Dixit <paritoshd@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>