boards: Add MMC CQE warning for Jetson Xavier
The following warning is observed around 25% of the time on boot for
Jetson Xavier ...

 WARNING KERN mmc0: running CQE recovery

This was first observed with Linux v5.17 and appears to be seen after
commit c2fee44399af ("arm64: tegra: Update SDMMC4 speeds for Tegra194")
was applied. There does not appear to be any obvious side-effect from
this and so add this to the allow list of warnings for now.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>