linux: Move sysfs into separate package
The module will be extended in upcoming patches, so move this into a
separate package for better code organization.

While at it, also make it possible to create a new sysfs device from
a specified parent and name pair. If no name is specified, this will
effectively create a copy of the parent.

Signed-off-by: Thierry Reding <treding@nvidia.com>