boards: Use new I2C sysfs classes
Replace existing I2C devices by the new I2C sysfs controller and device
classes, for a more concise representation.

Note that the Tegra234 changes also move the device creation into the
SoC class and references the per-SoC devices from the board files. This
should help reduce duplication down the line. The idea is that going
forward, all the devices that are available on a SoC should be defined
in the SoC files and then added to the list of devices for each board.
The list of devices per SoC is a representation of which devices exist,
while the list of devices per board designates devices that are active
for that board.

Since the old I2C controller and device classes are no longer used they
can be removed.

Signed-off-by: Thierry Reding <treding@nvidia.com>