tests: Add a host-side sanity test
This test is meant to be run on the host side and perform various sanity
checks on the code base. At this point in time it will iterate over all
the defined SoCs and boards and print some information about them. This
is sufficient to catch some easy to make errors that would otherwise
require them to be run on a test device.

Signed-off-by: Thierry Reding <treding@nvidia.com>