tests/boot: Improve handling of unspecified driver bindings
If a device isn't expected to be bound to a driver, by default we use a
"None" value. This isn't currently handled correctly in the checks and
causes failures at runtime.

Handle this properly by making anything that's not already a list into a
list, rather than just strings. This way, a "None" value will simply not
match any driver that might have bound to the device.

Signed-off-by: Thierry Reding <treding@nvidia.com>