boards: Remove snd-soc-dummy for Jetson AGX Orin
In Linux v6.15, commit 18abb3797f1c ("ASoC: soc-utils: Transition to the
faux device interface") no longer registers the 'snd-soc-dummy' as a
platform driver. Although we could check for the presence of this on
specific kernel versions, we don't really care about this device and
given that we do check for the presence of the actual HDA and APE
soundcards, let's remove this from the list of devices for Jetson AGX
Orin.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>