boards: Remove dummy regulator for Jetson AGX Orin
Commit dcd2a9a5550e ("regulator: dummy: convert to use the faux device
interface") has been added to the -next kernel and this causes the
devices test on Jetson AGX Orin to fail ...

 boot: devices:   device: /sys/bus/platform/devices/reg-dummy ... failed

Dummy regulator devices are used when regulators are not found but these
are not real devices. Therefore, just remove this device from the list
of devices for Jetson AGX Orin.

Signed-off-by: Thierry Reding <treding@nvidia.com>