boards: Allow BPMP sync_state() warnings
In Linux v6.17, commit f66c65686abd ("pmdomain: core: Default to
use of_genpd_sync_state() for genpd providers") assigns a
'->sync_state()' callback because none is assigned. This causes warnings
such as the following to be seen on platforms that support the BPMP ...

 WARNING KERN tegra-bpmp bpmp: sync_state() pending due to 17000000.gpu
 WARNING KERN tegra-bpmp bpmp: sync_state() pending due to 3960000.cec
 WARNING KERN tegra-bpmp bpmp: sync_state() pending due to 15380000.nvjpg
 WARNING KERN tegra-bpmp bpmp: sync_state() pending due to 154c0000.nvenc
 WARNING KERN tegra-bpmp bpmp: sync_state() pending due to 15a80000.nvenc

The above warnings occur because these devices have not been probed.
Apart from the CEC device, there are no upstream drivers for the above
drivers and so these are expected. Therefore, add these sync_state()
warnings observed on Tegra186 and Tegra194 devices to the allowed list.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>